Optimize Search commands that don't affect matches
Moving up and down and making a selection do not change the list of
matches, so we can eliminate needless computation from these actions
with a safe way to pass along precomputed matches to a new Search
instance.

This makes a big difference in performance when the list of choices is
large (> 1k).