Rust implementation of score
The score algorithm is implemented in Rust and exposes a `ext_score`
function. Build steps and testing instructions are in `score/README.md`.

The selecta script was modified to integrate with the `ext_score`
function by changing the `Score` class into a Module.

The use of `Fiddle::Importer` changes the minimum ruby requirement to
2.0.0.