draw query line last
When the query contains multi-character glyphs, this ensures that the
cursor is in the right place. That happens because the cursor is
naturally left at the end of the query, not through any kind of
intelligent computation. As far as I know, there's no way to compute the
number of glyphs in a Ruby string without including third-party
libraries.