Fix specs broken by replacing the bash and exec;
In the most recent PR (#107) a change was made to use 'env -S', however this
prevented the specs from running with 'bundle exec rspec'. This was due to
'split' in 'spec_helper.rb' no longer finding the correct string.

In this commit, the string is updated, allowing split to work correctly and the
tests to run.